Remove most of unneeded bakrep related code
