+++ src/components/PhylogenyTree.vue
@@ -11,12 +11,12 @@
 </template>
 
 <script setup lang="ts">
-import type { GtdbtkResult } from "@/model/GtdbtkResult";
+import type { SorfdbEntry } from "@/model/SorfdbSearchResult";
 import type { PropType } from "vue";
 import { computed } from "vue";
 
 const props = defineProps({
-  gtdb: { type: Object as PropType<GtdbtkResult>, default: undefined },
+  entry: { type: Object as PropType<SorfdbEntry>, required: true },
 });
 
 const mapping: Record<string, string> = {
@@ -45,18 +45,16 @@ function toEntry(level: string, value: string): TreeEntry {
 }
 
 const entries = computed(() => {
-  if (props.gtdb) {
-    const c = props.gtdb.classification;
-    return [
-      toEntry("domain", c.domain ? c.domain : ""),
-      toEntry("phylum", c.phylum ? c.phylum : ""),
-      toEntry("class", c.class ? c.class : ""),
-      toEntry("order", c.order ? c.order : ""),
-      toEntry("family", c.family ? c.family : ""),
-      toEntry("genus", c.genus ? c.genus : ""),
-      toEntry("species", c.species ? c.species : ""),
-    ];
-  }
-  return [];
+  const c = props.entry;
+  return [
+    toEntry("domain", "Bacteria"),
+    toEntry("phylum", c.phylum ? c.phylum : ""),
+    toEntry("class", c.class ? c.class : ""),
+    toEntry("order", c.order ? c.order : ""),
+    toEntry("family", c.family ? c.family : ""),
+    toEntry("genus", c.genus ? c.genus : ""),
+    toEntry("species", c.species ? c.species : ""),
+    toEntry("strain", c.strain ? c.strain : ""),
+  ];
 });
 </script>
