+++ src/views/browse/ResultTable.vue
@@ -1,17 +1,21 @@
 <script setup lang="ts">
-import type { BakrepSearchResultEntry } from "@/model/BakrepSearchResult";
+import type { SorfdbEntry } from "@/model/SorfdbSearchResult";
 import { computed, type PropType } from "vue";
 import type { SortDirection, SortOption } from "@/model/Search";
 import SortSymbol from "./SortSymbol.vue";
-
+import { type Option } from "@/components/CheckboxOption";
 const props = defineProps({
   entries: {
-    type: Array as PropType<BakrepSearchResultEntry[]>,
+    type: Array as PropType<SorfdbEntry[]>,
     default: () => [],
   },
   ordering: {
     type: Array as PropType<SortOption[]>,
     default: () => [],
+  },
+  visibleColumns: {
+    type: Array as PropType<Option[]>,
+    required: true,
   },
 });
 
@@ -19,45 +23,14 @@ const emit = defineEmits<{
   (e: "update:ordering", key: string, direction: SortDirection | null): void;
 }>();
 
-function gc(entry: BakrepSearchResultEntry): string {
-  if (!entry.bakta) return "?";
-  return (entry.bakta.stats.gc * 100).toFixed(2) + " %";
-}
-function contigs(entry: BakrepSearchResultEntry): string {
-  if (!entry.bakta) return "?";
-  return "" + entry.bakta.stats.no_sequences;
-}
-function genomeSize(entry: BakrepSearchResultEntry): string {
-  if (!entry.bakta) return "?";
-  return (
-    (Math.round(entry.bakta.stats.size / 10000) / 100).toLocaleString("en") +
-    " Mbp"
-  );
-}
-function species(entry: BakrepSearchResultEntry): string {
-  if (!entry.gtdbtk || !entry.gtdbtk.classification.species) return "?";
-  return entry.gtdbtk.classification.species;
-}
-function sequenceType(entry: BakrepSearchResultEntry): string {
-  if (!entry.mlst) return "?";
-  return entry.mlst.sequence_type;
-}
-function completeness(entry: BakrepSearchResultEntry): string {
-  if (!entry.checkm2) return "?";
-  return entry.checkm2.quality.completeness + "";
-}
-function contamination(entry: BakrepSearchResultEntry): string {
-  if (!entry.checkm2) return "?";
-  return entry.checkm2.quality.contamination + "";
-}
+// function gc(entry: BakrepSearchResultEntry): string {
+//   if (!entry.bakta) return "?";
+//   return (entry.bakta.stats.gc * 100).toFixed(2) + " %";
+// }
 
 function passOrdering(sortkey: string, newdirection: SortDirection | null) {
   emit("update:ordering", sortkey, newdirection);
 }
-
-const showFeatures = computed(() =>
-  props.entries.some((x) => x.bakta && x.bakta.features !== undefined),
-);
 </script>
 
 <template>
@@ -72,63 +45,6 @@ const showFeatures = computed(() =>
             @update:ordering="passOrdering"
           />
         </th>
-        <th scope="col">
-          GC
-          <SortSymbol
-            :ordering="ordering"
-            sortkey="bakta.stats.gc"
-            @update:ordering="passOrdering"
-          />
-        </th>
-        <th scope="col">
-          Contigs
-          <SortSymbol
-            :ordering="ordering"
-            sortkey="bakta.stats.no_sequences"
-            @update:ordering="passOrdering"
-          />
-        </th>
-        <th scope="col">
-          Genome Size
-          <SortSymbol
-            :ordering="ordering"
-            sortkey="bakta.stats.size"
-            @update:ordering="passOrdering"
-          />
-        </th>
-        <th scope="col">
-          Species
-          <SortSymbol
-            :ordering="ordering"
-            sortkey="gtdbtk.classification.species"
-            @update:ordering="passOrdering"
-          />
-        </th>
-        <th scope="col">
-          ST Type
-          <SortSymbol
-            :ordering="ordering"
-            sortkey="mlst.sequence_type"
-            @update:ordering="passOrdering"
-          />
-        </th>
-        <th scope="col">
-          Completeness
-          <SortSymbol
-            :ordering="ordering"
-            sortkey="checkm2.quality.completeness"
-            @update:ordering="passOrdering"
-          />
-        </th>
-        <th scope="col">
-          Contamination
-          <SortSymbol
-            :ordering="ordering"
-            sortkey="checkm2.quality.contamination"
-            @update:ordering="passOrdering"
-          />
-        </th>
-        <th v-if="showFeatures">Features</th>
       </tr>
     </thead>
     <tbody>
@@ -144,22 +60,6 @@ const showFeatures = computed(() =>
               {{ entry.id }}
             </router-link>
           </td>
-          <td>{{ gc(entry) }}</td>
-          <td>{{ contigs(entry) }}</td>
-          <td>
-            {{ genomeSize(entry) }}
-          </td>
-          <td>{{ species(entry) }}</td>
-          <td class="text-nowrap">{{ sequenceType(entry) }}</td>
-          <td>{{ completeness(entry) }} %</td>
-          <td>{{ contamination(entry) }} %</td>
-          <td v-if="showFeatures">
-            <ul class="nostyle" v-if="entry.bakta && entry.bakta.features">
-              <li v-for="(f, idx) of entry.bakta.features" :key="idx">
-                {{ f.gene }} - {{ f.product }}
-              </li>
-            </ul>
-          </td>
         </tr>
       </template>
     </tbody>
@@ -171,3 +71,4 @@ const showFeatures = computed(() =>
   cursor: pointer;
 }
 </style>
+@/model/SorfdbSearchResult
